Moving styles and functions out of component
