+++ src/models/modelTests/MyAccount.test.ts
@@ -76,7 +76,7 @@ describe("MyAccountModel", () => {
         {
           patron: "2772226",
           id: "48636910",
-          pickupByDate: "2024-02-15T09:00:00Z",
+          pickupByDate: "February 15, 2024",
           canFreeze: false,
           frozen: false,
           status: "Requested item ready for pickup.",
@@ -93,7 +93,7 @@ describe("MyAccountModel", () => {
           id: "65060571",
           callNumber: "972.93 D",
           barcode: "33333432264691",
-          dueDate: "2024-02-09T09:00:00Z",
+          dueDate: "February 9, 2024",
           patron: "2772226",
           title: "The Dominican Republic reader : history, culture, politics",
           isResearch: false,
@@ -104,7 +104,7 @@ describe("MyAccountModel", () => {
           id: "65060570",
           callNumber: "Spa FIC ALVAREZ",
           barcode: "33333455520789",
-          dueDate: "2024-02-09T09:00:00Z",
+          dueDate: "February 9, 2024",
           patron: "2772226",
           title: "En el tiempo de las mariposas",
           isResearch: false,
